refactor: Clean up mapper_handler
