+++ ahd2fhir/utils/mapper_handler.py
@@ -1,6 +1,10 @@
+import dataclasses
+from typing import Callable
+
 import structlog
 from fhir.resources.resource import Resource
 
+from ahd2fhir.config import Settings
 from ahd2fhir.mappers.ahd_to_condition import AHD_TYPE_DIAGNOSIS, get_fhir_condition
 from ahd2fhir.mappers.ahd_to_medication_statement import (
     AHD_TYPE_MEDICATION,
@@ -11,10 +15,30 @@ from ahd2fhir.mappers.ahd_to_observation_smkstat import (
     UKLFR_TYPE_SMKSTAT,
     get_fhir_resources,
 )
-from ahd2fhir.utils.custom_mappers import MapperBase
 from ahd2fhir.utils.device_builder import AHD_TYPE_DOCUMENT_ANNOTATION, build_device
 
 log = structlog.get_logger()
+
+
+@dataclasses.dataclass
+class MapperBase:
+    name: str
+    config: Settings
+    ahd_type: str
+    mapper_function: Callable
+    deduplicate_function: Callable = lambda x: x
+
+    def get_resources(self, ahd_response_entry, doc_ref):
+        return self.mapper_function(ahd_response_entry, doc_ref)
+
+    def deduplicate_resources(self, resources):
+        return self.deduplicate_function(resources)
+
+    def enabled(self) -> bool:
+        return True if self.ahd_type in self.config.mappers_enabled else False
+
+    def __repr__(self):
+        return self.name
 
 
 class MapperHandler:
@@ -45,19 +69,6 @@ class MapperHandler:
                 mapper_function=get_fhir_medication_statement,
                 deduplicate_function=deduplicate_resources,
             ),
-            MapperBase(
-                "DeviceMapper", config, AHD_TYPE_DOCUMENT_ANNOTATION, build_device
-            ),
-            MapperBase(
-                "ConditionMapper", config, AHD_TYPE_DIAGNOSIS, get_fhir_condition
-            ),
-            MapperBase(
-                "MedicationMapper",
-                config,
-                AHD_TYPE_MEDICATION,
-                get_fhir_medication_statement,
-                deduplicate_resources,
-            ),
         ]
         log.info(f"Enabled mappers: {[m for m in self.mappers if m.enabled()]}")
 
@@ -74,7 +85,6 @@ class MapperHandler:
                             mapper_results.append(results)
 
             mapper_results = mapper.deduplicate_resources(mapper_results)
-            # print(mapper, mapper_results)
             total_results.extend(mapper_results)
         print([[type(r)] for r in total_results])
         return total_results
