+++ tests/test_resource_handler.py
@@ -7,7 +7,9 @@ from fhir.resources.bundle import Bundle, BundleEntry
 from fhir.resources.composition import Composition
 from fhir.resources.documentreference import DocumentReferenceContent
 
+from ahd2fhir.utils.mapper_handler import MapperHandler
 from ahd2fhir.utils.resource_handler import ResourceHandler
+from tests.test_main import get_settings_override
 from tests.utils import get_empty_document_reference
 
 
@@ -25,19 +27,37 @@ class MockPipeline:
         return self.response
 
 
-def test_handle_bundle_with_empty_bundle_should_return_empty_result():
-    bundle = Bundle(**{"id": "test", "type": "transaction", "entry": []})
-    resource_handler = ResourceHandler(averbis_pipeline=MockPipeline())
+@pytest.fixture
+def resource_handler():
+    averbis_pipeline = MockPipeline()
+    mapper_handler = MapperHandler(get_settings_override())
+    return ResourceHandler(averbis_pipeline, mapper_handler)
 
+
+@pytest.fixture
+def resource_handler_with_mock_response():
+    ahd_response = []
+    with open("tests/resources/ahd/payload_creating_duplicate_medication.json") as file:
+        ahd_response = json.load(file)
+
+    averbis_pipeline = MockPipeline(response=ahd_response)
+    mapper_handler = MapperHandler(get_settings_override())
+
+    return ResourceHandler(averbis_pipeline, mapper_handler)
+
+
+def test_handle_bundle_with_empty_bundle_should_return_empty_result(resource_handler):
+    bundle = Bundle(**{"id": "test", "type": "transaction", "entry": []})
     bundle = resource_handler.handle_bundle(bundle)
 
     assert len(bundle.entry) == 0
 
 
-def test_handle_bundle_with_document_without_content_should_raise_error():
+def test_handle_bundle_with_document_without_content_should_raise_error(
+    resource_handler,
+):
     entry = BundleEntry(**{"resource": get_empty_document_reference()})
     bundle = Bundle(**{"id": "test", "type": "transaction", "entry": [entry]})
-    resource_handler = ResourceHandler(averbis_pipeline=MockPipeline())
 
     with pytest.raises(ValueError):
         bundle = resource_handler.handle_bundle(bundle)
@@ -45,7 +65,7 @@ def test_handle_bundle_with_document_without_content_should_raise_error():
 
 # TODO: should a document resulting in no annotations
 #       create resources within the bundle?
-def test_handle_bundle_with_content_should_create_composition():
+def test_handle_bundle_with_content_should_create_composition(resource_handler):
     doc = get_empty_document_reference()
     doc.content[0] = DocumentReferenceContent(
         **{
@@ -56,7 +76,6 @@ def test_handle_bundle_with_content_should_create_composition():
     )
     entry = BundleEntry(**{"resource": doc})
     bundle = Bundle(**{"id": "test", "type": "transaction", "entry": [entry]})
-    resource_handler = ResourceHandler(averbis_pipeline=MockPipeline())
 
     bundle = resource_handler.handle_bundle(bundle)
 
@@ -64,7 +83,9 @@ def test_handle_bundle_with_content_should_create_composition():
     assert len(bundle.entry) == 1
 
 
-def test_get_fhir_medication_should_only_create_unique_bundle_entries():
+def test_get_fhir_medication_should_only_create_unique_bundle_entries(
+    resource_handler_with_mock_response,
+):
     doc = get_empty_document_reference()
     doc.content[0] = DocumentReferenceContent(
         **{
@@ -74,15 +95,7 @@ def test_get_fhir_medication_should_only_create_unique_bundle_entries():
         }
     )
 
-    ahd_response = []
-    with open("tests/resources/ahd/payload_creating_duplicate_medication.json") as file:
-        ahd_response = json.load(file)
-
-    resource_handler = ResourceHandler(
-        averbis_pipeline=MockPipeline(response=ahd_response)
-    )
-
-    bundle = resource_handler.handle_documents([doc])
+    bundle = resource_handler_with_mock_response.handle_documents([doc])
 
     full_urls = [entry.fullUrl for entry in bundle.entry]
     unique_full_urls = set(full_urls)
